requirements.txt
# --- Core Python ---
pydantic
python-dotenv
tqdm
watchdog

# --- Backend / API ---
fastapi
uvicorn
flask
flask-cors
flask-socketio
python-socketio
eventlet

# --- RAG / Embeddings / Vector DB ---
faiss-cpu
chromadb
numpy
pandas
nltk
rank-bm25

# --- OpenAI / External APIs ---
openai
requests

# --- PDF / File processing ---
pypdf2
pdfplumber

# --- Audio / STT / TTS ---
pydub
soundfile
scipy
faster-whisper
python-multipart    # pour upload audio
ffmpeg-python       # recommandé si conversion audio
piper-tts

# --- Redis / Message Queue ---
redis
hiredis        # accélération optionnelle mais recommandée

# --- Web ---
jinja2

# --- Tests ---
pytest

# --- Utils ---
python-dateutil
